pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privateer_ad"
version = "0.1.0"
description = "Anomaly Detection with Differential Privacy"
authors = [
    {name = "Giorgos Livanos", email = "livanos.g@gmail.com"}
]
license = "Apache-2.0"
dependencies = [
    "dash==2.18.2",
    "dash_bootstrap_components==1.7.1",
    "fire==0.7.0",
    "flwr[simulation]==1.16.0",
    "flwr-datasets==0.5.0",
    "hatchling==1.27.0",
    "matplotlib==3.10.1",
    "mlflow==2.20.4",
    "numpy==1.26.4",
    "optuna==4.2.1",
    "opacus==1.5.3",
    "pandas==2.2.3",
    "plotly==6.0.0",
    "python-dotenv==1.0.1",
    "pytorch-forecasting==1.3.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "scikit-learn==1.6.1",
    "seaborn==0.13.2",
    # For AMD GPU compatible with ROCm=6.2.4
    #"torch @ https://download.pytorch.org/whl/rocm6.2.4/torch-2.6.0%2Brocm6.2.4-cp311-cp311-manylinux_2_28_x86_64.whl#sha256=e735d8c9bd19b7ed43274b921b23ca91887df78aa90cee7bfe295342572e4fc7",
    # For CPU-only system
    #"torch @ https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.5.1%2Bcpu.cxx11.abi-cp311-cp311-linux_x86_64.whl#sha256=05a962994d5e64832e6aff275a41cb12ead0a38632c4806b84de1b4f118d1a3b",
    # For CUDA-enabled devices
    "torch==2.6.0",
    "torchinfo==1.8.0",
    "tqdm==4.67.1"
]

[project.scripts]
train-ad = "privateer_ad.train.train:main"
eval-ad = "privateer_ad.evaluate.evaluate:main"
predict-ad = "privateer_ad.predict.predict:main"
autotune-ad = "privateer_ad.train.autotune:main"
server-app = "privateer_ad.fl.server_app:app"
client-app = "privateer_ad.fl.client_app:app"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.flwr.app]
publisher = "gliv"

[tool.flwr.app.components]
serverapp = "privateer_ad.fl.server_app:app"
clientapp = "privateer_ad.fl.client_app:app"

[tool.flwr.app.config]  # Run Config
num-server-rounds = 200
epochs = 5
fraction-fit =1.0
n_clients=3
run-name = "federated_learning"

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 3
options.backend.client-resources.num-cpus = 8 # Each ClientApp will get assigned n CPU core
options.backend.client-resources.num-gpus = 0.33 # Each ClientApp will get n * 100% of each available GPU VRAM
options.backend.init_args.num_cpus = 24 # Only expose N CPU to the simulation
options.backend.init_args.num_gpus = 1 # Expose N GPUs to the simulation
options.backend.max_restarts = 2
options.backend.max_task_retries = 2

[tool.flwr.federations.local-deployment]
address = "localhost:9099"
insecure = true
